refactor: 무거운 파일 분리
