+++ frontend/src/01_widgets/chrome/ui/brand.ts
@@ -1,0 +1,11 @@
+export const navLinks = [
+  { href: "/?tab=trending", label: "Trending" },
+  { href: "/explore", label: "Explore" },
+  { href: "/topics", label: "Topics" },
+] as const;
+
+export const logoWordmarkClassName =
+  "text-[24px] font-bold leading-none [font-family:Georgia,serif]";
+
+export const logoMarkClassName =
+  "font-bold leading-none [font-family:Georgia,serif]";
