+++ frontend/src/01_widgets/home-feed/ui/index.ts
@@ -1,0 +1,1 @@
+export { HomeFeed } from "./HomeFeed";
