+++ frontend/src/01_widgets/post/ui/index.ts
@@ -1,0 +1,4 @@
+export { PostArticle } from "./PostArticle";
+export { PostSuggests } from "./PostSuggests";
+export { PostTabs } from "./PostTabs";
+export { SuggestionDetail } from "./SuggestionDetail";
