+++ frontend/src/01_widgets/profile/ui/index.ts
@@ -1,0 +1,1 @@
+export { ProfileView } from "./ProfileView";
