+++ frontend/src/01_widgets/write/ui/index.ts
@@ -1,0 +1,1 @@
+export { WriteView } from "./WriteView";
