+++ frontend/src/02_features/auth/ui/index.ts
@@ -1,0 +1,1 @@
+export { GuestActions } from "./GuestActions";
