+++ frontend/src/02_features/discussion-composer/ui/index.ts
@@ -1,0 +1,1 @@
+export { DiscussionComposer } from "./DiscussionComposer";
