+++ frontend/src/03_entities/post/model/index.ts
@@ -1,0 +1,15 @@
+export {
+  getOpenLogPostEntry,
+  getOpenLogSuggestion,
+  openLogContributors,
+} from "./openLogPostData";
+export type {
+  OpenLogContributor,
+  OpenLogDiffRow,
+  OpenLogDiscussionComment,
+  OpenLogPost,
+  OpenLogPostEntry,
+  OpenLogReviewer,
+  OpenLogSuggestion,
+  OpenLogSuggestionComment,
+} from "./openLogPostData";
