+++ frontend/src/04_shared/lib/markdown/index.ts
@@ -1,0 +1,4 @@
+export {
+  formatSelection,
+  type ToolbarAction,
+} from "./openLogMarkdownFormatting";
