+++ frontend/src/04_shared/ui/icons/index.ts
@@ -1,0 +1,1 @@
+export { GitPullRequestIcon } from "./GitPullRequestIcon";
