+++ frontend/src/04_shared/ui/markdown/index.ts
@@ -1,0 +1,2 @@
+export { MarkdownContent } from "./MarkdownContent";
+export { MarkdownToolbar } from "./MarkdownToolbar";
